added browse evenets to the menu and extracted the header
