--- app_frontend/app/src/pages/HomePage.jsx
@@ -1,7 +1,0 @@
-import React from 'react'
-
-export default function HomePage() {
-  return (
-    <div>HomePage</div>
-  )
-}
