+++ package.json
@@ -1,6 +1,7 @@
 {
   "dependencies": {
     "express": "^4.21.2",
-    "pg": "^8.14.1"
+    "pg": "^8.14.1",
+    "react-router-dom": "^7.4.0"
   }
 }
